third_party/blink/web_tests/platform/generic/storage/websql/execute-sql-rowsAffected-expected.txt

This test tests that SQLResultSet.rowsAffected attribute is correct in success callback for executeSql().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rowsAffected is 1
PASS rowsAffected is 1
PASS rowsAffected is 2
PASS rowsAffected is 0
PASS rowsAffected is 2
PASS rowsAffected is 0
PASS rowsAffected is 0
PASS successfullyParsed is true

TEST COMPLETE
